Recognise styled-components as React Components
Currently styled-components are not being recognised as React components since they are an object wrapper around a React component.

The isReactComponent util function will now also look at the styledComponentId property since it's not present in the vue-styled-components implementation (so it won't mistakenly try to render a Vue styled-component as a React component).